// frontend/src/types/api.enum.ts
export enum ApiEnum {
    DASHBOARD = '/',
    LOGIN = '/login',
    SIGNUP = '/signup',
    LOGOUT = '/logout',
    REFRESH = '/refresh',
    INCOME = '/income',
    CREATE_INCOME = '/income/create',
    INCOME_EDIT = '/income/edit',
    DELETE_INCOME = '/income/delete',
    EXPENSES = '/expenses',
    CREATE_EXPENSES = '/expenses/create',
    EDIT_EXPENSES = '/expenses/edit',
    DELETE_EXPENSE = '/expense/delete',
    ROUTE_OPERATIONS = '/operations',
    CREATE_OPERATION = '/create-operation',
    EDIT_OPERATION = '/edit-operation',
    BALANCE = '/balance',
    GET_CATEGORIES_EXPENSE = '/categories/expense',
    GET_CATEGORIES_INCOME = '/categories/income',
    OPERATIONS_DELETE = '/operations/delete',
    OPERATIONS = '/operations'
}